checkin distance

+++ app/src/main/java/nctu/cs/cgv/itour/object/Mesh.java
@@ -10,19 +10,19 @@ import java.util.StringTokenizer;
 
 public class Mesh {
     // world coordinates bounding box
-    public double minLat = 0;
-    public double minLon = 0;
-    public double maxLat = 0;
-    public double maxLon = 0;
+    public float minLat = 0;
+    public float minLon = 0;
+    public float maxLat = 0;
+    public float maxLon = 0;
     //real-osm-map size
-    public double mapWidth = 0;
-    public double mapHeight = 0;
+    public float mapWidth = 0;
+    public float mapHeight = 0;
     // element count
     int vertexNumber;
     int faceNumber;
     int lineNumber;
     // x- and y- coordinates of nodeMap
-    double[][] vertices;
+    float[][] vertices;
 
     // face connectivity
     int[][] faces;
@@ -50,7 +50,7 @@ public class Mesh {
             lineNumber = Integer.valueOf(stringTokenizer.nextToken());
 
             // get arrayLists
-            vertices = new double[vertexNumber][2];
+            vertices = new float[vertexNumber][2];
             faces = new int[faceNumber][3];
 
             // read vertex positions
@@ -58,8 +58,8 @@ public class Mesh {
                 nextLine = bufferedReader.readLine();
 
                 stringTokenizer = new StringTokenizer(nextLine);
-                double x = Double.valueOf(stringTokenizer.nextToken());
-                double y = Double.valueOf(stringTokenizer.nextToken());
+                float x = Float.valueOf(stringTokenizer.nextToken());
+                float y = Float.valueOf(stringTokenizer.nextToken());
 
                 vertices[vIter][0] = x;
                 vertices[vIter][1] = y;
@@ -99,22 +99,22 @@ public class Mesh {
 
             // in the order of minlat -> minlon -> maxlat -> maxlon ->mapWidth ->mapHeight
             nextLine = bufferedReader.readLine();
-            minLat = Double.valueOf(nextLine);
+            minLat = Float.valueOf(nextLine);
 
             nextLine = bufferedReader.readLine();
-            minLon = Double.valueOf(nextLine);
+            minLon = Float.valueOf(nextLine);
 
             nextLine = bufferedReader.readLine();
-            maxLat = Double.valueOf(nextLine);
+            maxLat = Float.valueOf(nextLine);
 
             nextLine = bufferedReader.readLine();
-            maxLon = Double.valueOf(nextLine);
+            maxLon = Float.valueOf(nextLine);
 
             nextLine = bufferedReader.readLine();
-            mapWidth = Double.valueOf(nextLine);
+            mapWidth = Float.valueOf(nextLine);
 
             nextLine = bufferedReader.readLine();
-            mapHeight = Double.valueOf(nextLine);
+            mapHeight = Float.valueOf(nextLine);
 
             bufferedReader.close();
 
@@ -123,7 +123,7 @@ public class Mesh {
         }
     }
 
-    public IdxWeights getPointInTriangleIdx(double px, double py) {
+    public IdxWeights getPointInTriangleIdx(float px, float py) {
 
         IdxWeights result = new IdxWeights();
 
@@ -132,16 +132,16 @@ public class Mesh {
             int id2 = faces[fIter][1];
             int id3 = faces[fIter][2];
 
-            double x1 = vertices[id1][0];
-            double x2 = vertices[id2][0];
-            double x3 = vertices[id3][0];
-            double y1 = vertices[id1][1];
-            double y2 = vertices[id2][1];
-            double y3 = vertices[id3][1];
+            float x1 = vertices[id1][0];
+            float x2 = vertices[id2][0];
+            float x3 = vertices[id3][0];
+            float y1 = vertices[id1][1];
+            float y2 = vertices[id2][1];
+            float y3 = vertices[id3][1];
 
-            double l1 = ((y2 - y3) * (px - x3) + (x3 - x2) * (py - y3)) / ((y2 - y3) * (x1 - x3) + (x3 - x2) * (y1 - y3));
-            double l2 = ((y3 - y1) * (px - x3) + (x1 - x3) * (py - y3)) / ((y2 - y3) * (x1 - x3) + (x3 - x2) * (y1 - y3));
-            double l3 = 1.0f - l1 - l2;
+            float l1 = ((y2 - y3) * (px - x3) + (x3 - x2) * (py - y3)) / ((y2 - y3) * (x1 - x3) + (x3 - x2) * (y1 - y3));
+            float l2 = ((y3 - y1) * (px - x3) + (x1 - x3) * (py - y3)) / ((y2 - y3) * (x1 - x3) + (x3 - x2) * (y1 - y3));
+            float l3 = 1.0f - l1 - l2;
 
             if (l1 < 0 && l1 > -0.0001) l1 = 0f;
             if (l2 < 0 && l2 > -0.0001) l2 = 0f;
@@ -159,16 +159,16 @@ public class Mesh {
         return result;
     }
 
-    public double[] interpolatePosition(IdxWeights idxWeights) {
+    public float[] interpolatePosition(IdxWeights idxWeights) {
         int id = idxWeights.idx;
         int triId1 = faces[id][0];
         int triId2 = faces[id][1];
         int triId3 = faces[id][2];
-        double[] weights = idxWeights.weights;
+        float[] weights = idxWeights.weights;
 
-        double x = weights[0] * vertices[triId1][0] + weights[1] * vertices[triId2][0] + weights[2] * vertices[triId3][0];
-        double y = weights[0] * vertices[triId1][1] + weights[1] * vertices[triId2][1] + weights[2] * vertices[triId3][1];
+        float x = weights[0] * vertices[triId1][0] + weights[1] * vertices[triId2][0] + weights[2] * vertices[triId3][0];
+        float y = weights[0] * vertices[triId1][1] + weights[1] * vertices[triId2][1] + weights[2] * vertices[triId3][1];
 
-        return new double[]{x, y};
+        return new float[]{x, y};
     }
 }
